LayoutTests/media/controls/forced-tracks-only-expected.txt

This verifies that the track menu is not visible when a file has only forced tracks.

This test only runs in DRT!


EVENT: canplaythrough
PASS: We are using the apple idiom
PASS: All text tracks accounted for
PASS: Text track 1 not visible
PASS: Text track 2 not visible
PASS: Menu name is 'Track Menu'
PASS: Menu does not exist

Testing finished.
